Most Recent TMS_Waveform update V4
This update increases the accuracy of the boost pulse, as well as adds a distinction between the 20 degree and 40 degree pulses. Pulse width may not be as accurate as desired and future iterations will work towards increasing accuracy. I am also including a 23us Boost pulse waveform that was recorded from a digital simulation. To use this waveform, simply select 'Load Pulse from File' when prompted by MATLAB and select the .mat file.